Attach space to index (bad) and attach indices to ScalarExpr (good)
See FIXME comment for why we should not attach the space to the index.

It is a good idea to give all ScalarExpr subclasses indices because it
means that we can treat binops the same as indexed tensors (apart from
the fact that the ordering of the indices does not matter for binops).